% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroboclient.R
\name{absPosition.robo}
\alias{absPosition.robo}
\title{Posição no sistema de referência abosluto}
\usage{
\method{absPosition}{robo}(dev)
}
\arguments{
\item{dev}{Um objeto de classe \code{robo} contendo informações sobre servidor XML-RPC}
}
\value{
Lista contendo as coordenadas x, y e z.
}
\description{
Retorna a posição do robô cartesiano no sistema de referência absoluto.
O sistema de referência absoluto é o que está na placa controladora do Robo
}
\examples{
absPosition(dev)

}
