% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroboclient.R
\name{homeX.robo}
\alias{homeX.robo}
\title{Executa o comando de home no eixo x}
\usage{
\method{homeX}{robo}(dev, s = "+")
}
\arguments{
\item{dev}{Um objeto de classe \code{robo} contendo informações sobre servidor XML-RPC}

\item{s}{Caracter \code{'+'} ou \code{'-'} com direção do home.}
}
\description{
Volta o robô para a origem do sistema do eixo x. Como sensores de fim de curso
estão em ambos os extremos do eixo, deve-se especificar em que direção se deseja ir.
\code{'+'} na direção de x positivo (em direção ao ventilador), \code{'-'} na direção
de x negativo (em direção à contração). Lembre que este movimento se dá em velocidade
bem baixa e recomenda-se realizar boa parte da movimentação com o comando \code{\link{move}}.
}
\examples{
homeX(dev, '-')

}
