% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroboclient.R
\name{homeY.robo}
\alias{homeY.robo}
\title{Executa o comando de home no eixo y}
\usage{
\method{homeY}{robo}(dev, s = "-")
}
\arguments{
\item{dev}{Um objeto de classe \code{robo} contendo informações sobre servidor XML-RPC}

\item{s}{Caracter \code{'+'} ou \code{'-'} com direção do home.}
}
\description{
Volta o robô para a origem do sistema do eixo y. Como sensores de fim de curso
estão em ambos os extremos do eixo, deve-se especificar em que direção se deseja ir.
\code{'+'} na direção de y positivo (lado da porta no túnel de vento), \code{'-'} na direção
de x negativo (lado das salas de controle do túnel de vento). Lembre que este movimento se dá em velocidade
bem baixa e recomenda-se realizar boa parte da movimentação com o comando \code{\link{move}}.
}
\examples{
homeY(dev, '-')

}
