% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroboclient.R
\name{homeZ.robo}
\alias{homeZ.robo}
\title{Executa o comando de home no eixo z}
\usage{
\method{homeZ}{robo}(dev, s = "+")
}
\arguments{
\item{dev}{Um objeto de classe \code{robo} contendo informações sobre servidor XML-RPC}

\item{s}{Argumento ignorado}
}
\description{
Volta o robô para a origem do sistema do eixo z. Como o eixo z tem apenas um sensor de fim de curso,
ao contrário das funções \code{\link{homeX}} e \code{\link{homeY}} não se necessita especificar a
orientação: o home é sempre para cima! Para manter compatibilidade, o argumento de direção existe
mas é ignorado.
}
\examples{
homeZ(dev)

}
