% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroboclient.R
\name{move.robo}
\alias{move.robo}
\title{Movimenta robô cartesiano}
\usage{
\method{move}{robo}(dev, x = "", y = "", z = "", a = FALSE,
  r = FALSE, sync = FALSE)
}
\arguments{
\item{dev}{Um objeto de classe \code{robo} contendo informações sobre servidor XML-RPC}

\item{x}{Número informando o quanto se deve movimentar no eixo x}

\item{y}{Número informando o quanto se deve movimentar no eixo y}

\item{z}{Número informando o quanto se deve movimentar no eixo z}

\item{a}{TRUE/FALSE, informa se o movimento se dá na referência absoluta ou relativa}

\item{r}{TRUE/FALSE, movimento relativo em relação ao ponto atual?}

\item{@sync}{TRUE/FALSE Se sincronizado, a função só retorna depois de executado o movimento}
}
\description{
Executa os movimentos do robô cartesiano
}
\examples{
move(dev, x=5, r=TRUE)

}
