% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroboclient.R
\name{position.robo}
\alias{position.robo}
\title{Posição do robô}
\usage{
\method{position}{robo}(dev)
}
\arguments{
\item{dev}{Um objeto de classe \code{robo} contendo informações sobre servidor XML-RPC}
}
\value{
Lista contendo as coordenadas x, y e z.
}
\description{
Retorna a posição do robô cartesiano no sistema de referência relativo.
O sistema de referência absoluto é o que está na placa controladora do Robo
}
\examples{
position(dev)

}
