% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroboclient.R
\name{rmove.robo}
\alias{rmove.robo}
\title{Movimento relativo do robô cartesiano}
\usage{
\method{rmove}{robo}(dev, x = "", y = "", z = "", sync = FALSE)
}
\arguments{
\item{dev}{Um objeto de classe \code{robo} contendo informações sobre servidor XML-RPC}

\item{x}{Número informando o quanto se deve movimentar no eixo x}

\item{y}{Número informando o quanto se deve movimentar no eixo y}

\item{z}{Número informando o quanto se deve movimentar no eixo z}

\item{@sync}{TRUE/FALSE Se sincronizado, a função só retorna depois de executado o movimento}
}
\description{
Executa os movimentos do robô cartesiano em relação à posição atual
}
\examples{
rmove(dev, x=5)

}
