% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroboclient.R
\name{rmoveY.robo}
\alias{rmoveY.robo}
\title{Movimento relativo do robô cartesiano no eixo y}
\usage{
\method{rmoveY}{robo}(dev, x, sync = FALSE)
}
\arguments{
\item{dev}{Um objeto de classe \code{robo} contendo informações sobre servidor XML-RPC}

\item{x}{Número informando o quanto se deve movimentar no eixo y}

\item{@sync}{TRUE/FALSE Se sincronizado, a função só retorna depois de executado o movimento}
}
\description{
Executa os movimentos do robô cartesiano em relação à posição atual ao longo do eixo y
}
\examples{
rmoveY(dev, 5)

}
