% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroboclient.R
\name{roboDev}
\alias{roboDev}
\title{Criar conexão com robô cartesiano}
\usage{
roboDev(url = "localhost", port = 9595)
}
\arguments{
\item{url}{String com URL do servidor XML-RPC}

\item{port}{Inteiro com o número da porta TCP/IP usado pelo XML-RPC}
}
\value{
Um objeto de classe \code{robo}
}
\description{
Cria um objeto de classe \code{robo} que controla a comunicação com
o servidor XML-RPC do robô cartesiano do túnel de vento do IPT.
}
\examples{
dev <- roboDev()

}
