% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroboclient.R
\name{setReference.robo}
\alias{setReference.robo}
\title{Cria novo sistema de referência}
\usage{
\method{setReference}{robo}(dev, xref = 0, yref = 0, zref = 0)
}
\arguments{
\item{dev}{Um objeto de classe \code{robo} contendo informações sobre servidor XML-RPC}

\item{xref}{Valor que a coordenada x atual toma.}

\item{yref}{Valor que a coordenada y atual toma.}

\item{zref}{Valor que a coordenada z atual toma.}
}
\description{
Configura um novo sistema de referência onde a posição atual do robô tem
valores \code{xref}, \code{yref}, \code{zref}. Este sistema de referência relativo
tem a mesma orientação que o sistema de placa controladora. O que muda é a origem.
}
\examples{
setReference(dev, yref=100)

}
