% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroboclient.R
\name{stopnow.robo}
\alias{stopnow.robo}
\title{Paramada imediata}
\usage{
\method{stopnow}{robo}(dev)
}
\arguments{
\item{dev}{Um objeto de classe \code{robo} contendo informações sobre servidor XML-RPC}
}
\description{
Envia comando para o robô cartesiano para qualquer movimento.
}
\examples{
moveX(dev, 100)
stopnow(dev)

}
