% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroboclient.R
\name{waitUntilDone.robo}
\alias{waitUntilDone.robo}
\title{Espere o movimento terminar}
\usage{
\method{waitUntilDone}{robo}(dev)
}
\arguments{
\item{dev}{Um objeto de classe \code{robo} contendo informações sobre servidor XML-RPC}
}
\description{
A maioria dos comandos de movimentação do robô são assíncronos. O comando é dado e o programa
volta para o programador (usuário) imediatamente enquanto o movimento se realiza. Várias funções
têm o argumento \code{sync} que torna o comando síncrono. Mas no caso geral, esta função vai
simplesmente esperar que qualquer movimento programado termine antes de retornar o controle.
}
\examples{
move(dev, x=-500, sync=FALSE)
waitUntilDone(dev)

}
